Update wrappingfit_ALL
Now this script is called ExtractingData, which only extracts the data, pulls in trialwise analysus from PDGraphingTR_K and creates a trialwise table as well as a shortTable for linear models. Also extracts the allratings_g which is used in the conditional plot of all three groups' ratings over the trials.